Major updates on Upload Book Page
Upload book page can now select or photo image to upload to the DB. After selecting an image, user can see a preview on their phone screen. The best width and Hight should be 300x300, we will fix the amount later. The Upload to DB part is not yet ready (half-done) since we are now trying to make use of GridFS, which will then make our image become a link in DB. For buttons, long press will show its whats it for.